requirements.txt
Flask==3.0.3
google-cloud-run==0.10.4
gunicorn==23.0.0

requests==2.32.4
structlog==25.1.0

google-auth==2.3.2
